Deprecate `Shared`'s `Collection` conformances (#26)
Collections are expected to be value types that can be copied around and
compared, and while the library does put in some effort to make this
work, there are too many unknowns with how SwiftUI's `ForEach` probes
its given collection to contend with, and so we should go back to
deprecating these conformances entirely in favor of explicit conversions
to an actual collection.

Fixes #24.